refactor: shared UI restructure (#508)
* refactor: create ui folder and move shared primitives

* refactor: move appbar to layout folder
