+++ apps/container/src/components/ui/ReviewDialog/ReviewDialog.vue
@@ -90,8 +90,8 @@ import { ElMessage } from 'element-plus';
 import { computed, PropType, ref, watch } from 'vue';
 
 import { CommentsList } from '@/components/CommentsList';
-import { FeedbackAlert } from '@/components/FeedbackAlert';
-import { PaperCard } from '@/components/PaperCard';
+import { FeedbackAlert } from '@/components/ui/FeedbackAlert';
+import { PaperCard } from '@/components/ui/PaperCard';
 import { capitalizeName } from '@/utils/capitalizeName';
 import { conceptsColor } from '@/utils/consts';
