+++ apps/container/src/views/Recovery/RecoveryView.vue
@@ -6,7 +6,7 @@ import { useField, useForm } from 'vee-validate';
 import { computed, ref, watch } from 'vue';
 import { useRouter } from 'vue-router';
 
-import { FeedbackAlert } from '@/components/FeedbackAlert';
+import { FeedbackAlert } from '@/components/ui/FeedbackAlert';
 
 import { recoverySchema } from './recoveryValidationSchema';
